+++ NEWS.md
@@ -6,6 +6,7 @@
   for arguments to `delta_options`
 * Add `time_term` to allow time-variable interaction (e.g., AR1 intercepts)
 * Adding overview and model-description vignettes
+* Add `simulate` S3 generic
 
 # tinyVAST 0.7.1